# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "North Landing Disc Return",
    "version": "0.1.0"
  },
  "paths": {
...
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Me",
        "operationId": "getMe",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/users/me/phones": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Add Phone",
        "operationId": "addPhone",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPhoneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneNumberOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/users/me/phones/verify": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Verify Phone",
        "operationId": "verifyPhone",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPhoneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneNumberOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/users/me/phones/{number}": {
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Remove Phone",
        "operationId": "removePhone",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/me/wishlist": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get My Wishlist",
        "operationId": "getMyWishlist",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DiscOut"
                  },
                  "type": "array",
                  "title": "Response Getmywishlist"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Add Wishlist Disc",
        "operationId": "addWishlistDisc",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WishlistDiscCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/users/me/discs": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get My Discs",
        "operationId": "getMyDiscs",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DiscOut"
                  },
                  "type": "array",
                  "title": "Response Getmydiscs"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/users/me/wishlist/{disc_id}": {
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Remove Wishlist Disc",
        "operationId": "removeWishlistDisc",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "disc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Disc Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddPhoneRequest": {
        "properties": {
          "number": {
            "type": "string",
            "title": "Number"
          }
        },
        "type": "object",
        "required": [
          "number"
        ],
        "title": "AddPhoneRequest"
      },
...
      "DiscOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "manufacturer": {
            "type": "string",
            "title": "Manufacturer"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "colors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Colors"
          },
          "owner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OwnerOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_clear": {
            "type": "boolean",
            "title": "Is Clear"
          },
          "input_date": {
            "type": "string",
            "format": "date",
            "title": "Input Date"
          },
          "is_found": {
            "type": "boolean",
            "title": "Is Found"
          },
          "is_returned": {
            "type": "boolean",
            "title": "Is Returned"
          },
          "final_notice_sent": {
            "type": "boolean",
            "title": "Final Notice Sent"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/DiscPhotoOut"
            },
            "type": "array",
            "title": "Photos",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "manufacturer",
          "name",
          "colors",
          "is_clear",
          "input_date",
          "is_found",
          "is_returned",
          "final_notice_sent",
          "created_at",
          "updated_at"
        ],
        "title": "DiscOut"
      },
...
      "DiscPhotoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "photo_path": {
            "type": "string",
            "title": "Photo Path"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order"
          }
        },
        "type": "object",
        "required": [
          "id",
          "photo_path",
          "sort_order"
        ],
        "title": "DiscPhotoOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OwnerOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "heads_up_sent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heads Up Sent At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_name",
          "last_name",
          "phone_number",
          "created_at",
          "name"
        ],
        "title": "OwnerOut"
      },
...
      "PhoneNumberOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "number": {
            "type": "string",
            "title": "Number"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "verified",
          "verified_at"
        ],
        "title": "PhoneNumberOut"
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin"
          },
          "phone_numbers": {
            "items": {
              "$ref": "#/components/schemas/PhoneNumberOut"
            },
            "type": "array",
            "title": "Phone Numbers",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "is_admin",
          "created_at"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyPhoneRequest": {
        "properties": {
          "number": {
            "type": "string",
            "title": "Number"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "number",
          "code"
        ],
        "title": "VerifyPhoneRequest"
      },
      "WishlistDiscCreate": {
        "properties": {
          "manufacturer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manufacturer"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "colors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Colors"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "owner_first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner First Name"
          },
          "owner_last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Last Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "phone_number"
        ],
        "title": "WishlistDiscCreate"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}